crud done

+++ client/src/components/admin/MainSection.vue
@@ -6,11 +6,11 @@
       </router-link>
       <router-view></router-view>
     </div>
-    <hr style="border:1px grey solid">
+    <hr style="border:1px grey solid" />
     <div class="list">
       <table>
-        <thead >
-          <tr >
+        <thead>
+          <tr>
             <th>Id</th>
             <th>Name</th>
             <th>Image Url</th>
@@ -22,7 +22,7 @@
           </tr>
         </thead>
         <tbody>
-          <tr v-for="item in products" :key="item.id">
+          <tr v-for="item in products" :key="item.id" :dataItem="item">
             <td>{{item.id}}</td>
             <td>{{item.name}}</td>
             <td>{{item.image_url}}</td>
@@ -31,9 +31,31 @@
             <td>{{item.category}}</td>
             <td>{{item.description}}</td>
             <td>
-              <b-button variant="info" style="padding:3px 5px;margin:3px 3px;">Edit</b-button>
-              <b-button variant="danger" style="padding:3px;margin:2px 3px;" @click="deleteItem(item)">Delete</b-button>
+              <b-button
+                v-b-modal.modal-center
+                variant="info"
+                style="padding:3px 5px;margin:3px 3px;"
+              >
+                <i class="fa fa-edit"></i>
+              </b-button>
+              <b-button
+                variant="danger"
+                style="padding:3px 5px;margin:3px 3px;"
+                @click="deleteItem(item)"
+              >
+                <i class="fa fa-trash"></i>
+              </b-button>
             </td>
+            <b-modal
+              id="modal-center"
+              centered
+              title="Edit Item"
+              ok-only
+              ok-variant="secondary"
+              ok-title="Cancel"
+            >
+              <EditForm :product="item"></EditForm>
+            </b-modal>
           </tr>
         </tbody>
       </table>
@@ -43,14 +65,17 @@
 
 <script>
 import AddProduct from "./AddProduct";
+import EditForm from "./EditForm";
 
 export default {
   name: "MainSection",
   components: {
-    AddProduct
+    AddProduct,
+    EditForm
   },
   data() {
     return {
+      dataItem: null
     };
   },
   computed: {
@@ -58,34 +83,48 @@ export default {
       return this.$store.state.products;
     }
   },
-  methods:{
-    deleteItem(item){
-      this.$store.dispatch('deleteData',item)
+  methods: {
+    deleteItem(item) {
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      }).then(result => {
+        if (result.value) {
+          this.$store.dispatch("deleteData", item);
+          Swal.fire("Deleted!", "Your file has been deleted.", "success");
+        }
+      });
     }
   }
 };
 </script>
 
 <style scoped>
-  #add-button {
-    text-align: center;
-    margin: 20px auto;
-  }
-  th{
-    background:rgb(235, 235, 235);
-    padding: 5px 30px;
-    border: 1px solid rgb(182, 182, 182);
-  }
-  #main-section{
-    text-align: center;
-    margin: auto;
-  }
-  td{
-    border: 1px solid rgb(182, 182, 182);
-  }
-  table{
-    margin :auto;
-    margin-bottom: 20px;
-  }
-  
+#add-button {
+  text-align: center;
+  margin: 20px auto;
+}
+th {
+  background: rgb(235, 235, 235);
+  padding: 5px 5px;
+  border: 1px solid rgb(182, 182, 182);
+}
+#main-section {
+  text-align: center;
+  margin: auto;
+}
+td {
+  border: 1px solid rgb(182, 182, 182);
+  padding: 3px;
+}
+table {
+  margin: auto;
+  margin-bottom: 20px;
+  width: 90%;
+}
 </style>
